+++ src/main/java/com/jteam/GroupProject/service/impl/UserServiceImpl.java
@@ -1,19 +1,23 @@
 package com.jteam.GroupProject.service.impl;
 
 
+import com.jteam.GroupProject.exceptions.NotFoundException;
 import com.jteam.GroupProject.model.User;
 import com.jteam.GroupProject.repository.UserRepository;
 import com.jteam.GroupProject.service.UserService;
+import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
-    public UserServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
+    private final Logger logger = LoggerFactory.getLogger(UserServiceImpl.class);
 
     /**
      * Создание и сохранение пользователя в бд
@@ -23,7 +27,10 @@ public class UserServiceImpl implements UserService {
      */
     @Override
     public User create(User user) {
-        return userRepository.save(user);
+
+        User savedUser = userRepository.save(user);
+        logger.info("User created: {}", savedUser);
+        return savedUser;
     }
 
     /**
@@ -34,7 +41,9 @@ public class UserServiceImpl implements UserService {
      */
     @Override
     public User getById(Long id) {
-        return userRepository.findAllById(id);
+        return userRepository.findByTelegramId(id)
+                .orElseThrow(() -> new NotFoundException("Пользователь не найден!"));
+
     }
 
     /**
@@ -45,7 +54,7 @@ public class UserServiceImpl implements UserService {
      */
     @Override
     public String getShelterById(Long id) {
-        return null;
+        return getById(id).getShelterType();
     }
 
     /**
@@ -64,7 +73,11 @@ public class UserServiceImpl implements UserService {
      */
     @Override
     public User update(User user) {
-        return userRepository.save(user);
+        User currentUser = getById(user.getTelegramId());
+        EntityUtils.copyNonNullFields(user, currentUser);
+        User updatedUser = userRepository.save(currentUser);
+        logger.info("User updated: {}", updatedUser);
+        return updatedUser;
     }
 
     /**
@@ -72,7 +85,9 @@ public class UserServiceImpl implements UserService {
      */
     @Override
     public void delete(User user) {
-        userRepository.delete(user);
+
+        userRepository.delete(getById(user.getTelegramId()));
+        logger.info("User deleted: {}", user);
     }
 
     /**
@@ -82,6 +97,11 @@ public class UserServiceImpl implements UserService {
      */
     @Override
     public void deleteById(Long id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            logger.info("User deleted by ID: {}", id);
+        } else {
+            logger.warn("User with ID {} not found, delete operation skipped.", id);
+        }
     }
 }
